PlotlyJSWrapper.add_background_traces! — Method

add_background_traces!(traces, bkgs; options)

Add background traces to traces

PlotlyJSWrapper.add_cms_label! — Method

add_cms_label!(annotations; options)

Add CMS label to annotations

PlotlyJSWrapper.add_data_traces! — Method

add_data_traces!(traces, data; options)

Add data traces to traces

PlotlyJSWrapper.add_lumi_label! — Method

add_lumi_label!(annotations; options)

Add luminosity and energy label to annotations

PlotlyJSWrapper.add_ratio_axes_traces! — Method

add_ratio_axes_traces!(traces; options)

Add ratio axes traces to traces

PlotlyJSWrapper.add_ratio_traces! — Method

add_ratio_traces!(traces, ratio; options)

Add data / MC ratio traces to traces

PlotlyJSWrapper.add_total_traces! — Method

add_total_traces!(traces, total; options)

Add total background traces to traces

PlotlyJSWrapper.build_hist1dtrace — Method

build_hist1dtrace(h; witherror=false, datastyle=false)

Create plotly trace representing Hist1D data. There are three different variations. if datastyle = true, then it returns a single scatter trace. if witherror = true, then it returns an array of 'scatter' traces. The first trace is the central value, while the following two traces are the up and down uncertainty bands. If all extra arguments are false, it returns a single bar histogram. (i.e. background style)

PlotlyJSWrapper.colors — Method

colors(i; opacity)

Custom colors

PlotlyJSWrapper.fit_bkg_to_data! — Method

fit_bkg_to_data!(bkgs, total, data_)

Scales bkgs and total to first data integral

PlotlyJSWrapper.get_ymax — Method

get_ymax(total, signals, data_)

Finds the maximum y point including errors

PlotlyJSWrapper.get_ymin — Method

get_ymax(total, signals, data_)

Finds the minimum y point including errors

PlotlyJSWrapper.get_yrange — Method

get_ymax(total, signals, data_)

Get range of ymin to ymax

PlotlyJSWrapper.plot_stack — Method

plot_stack(; backgrounds, signals=[], data=[], options::Dict{Symbol, Any}=default_options)

Create HEP-style stacked plot with Data/MC ratio in the bottom panel. backgrounds, signals, and data, are Arrays of Hist1D.